Only create the archive file when it is done
Because GitLab caches repository archive files on disk it is important
to avoid leaving a corrupt archive when a Unicorn worker gets
interrupted. If we do not watch out for this, the next user to download
the same revision as an archive gets the unfinished file from the
interrupted worker.

This change uses a (hopefully atomic) rename to put the archive file
in its expected location.